+++ src/common/types.ts
@@ -101,4 +101,15 @@ export interface ParsedCDEvent {
     deploymentHistoryLink: string;
     dockerImg: string;
 }
+export enum EVENT_TYPE {
+    Trigger = 1,
+    Success = 2,
+    Fail = 3,
+    Approval = 4,
+    ConfigApproval = 5,
+    Blocked = 6,
+    ImagePromotion = 7,
+    ImageScan = 8
+
+}
