--- src/common/utils.ts
@@ -1,12 +1,0 @@
-export enum EVENT_TYPE  {
-    Trigger= 1, 
-    Success= 2,
-    Fail= 3,
-    Approval= 4,
-    ConfigApproval= 5,
-    Blocked=6,
-    ImagePromotion=7,
-    ImageScan=8
-    
-};
-
